Extract EditPostAction
