+++ config/routes.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 use Sihae\Actions\ArchivedPostsAction;
 use Sihae\Actions\CreatePostAction;
 use Sihae\Actions\DeletePostAction;
+use Sihae\Actions\EditPostAction;
 use Sihae\Actions\EditPostFormAction;
 use Sihae\Actions\LoginAction;
 use Sihae\Actions\LoginFormAction;
@@ -31,7 +32,7 @@ return static function (App $app): void {
 
         $group->group('', function (RouteCollectorProxy $group): void {
             $group->get('/edit/{slug:[a-zA-Z\d\s\-_\-]+}', EditPostFormAction::class);
-            $group->post('/edit/{slug:[a-zA-Z\d\s\-_\-]+}', PostController::class . ':update');
+            $group->post('/edit/{slug:[a-zA-Z\d\s\-_\-]+}', EditPostAction::class);
             // TODO this should be a POST request
             $group->get('/delete/{slug:[a-zA-Z\d\s\-_\-]+}', DeletePostAction::class);
         })->add(PostLocator::class);
